requirements.txt
streamlit
pymupdf
python-docx
spacy
pandas
numpy
matplotlib
reportlab
